Reorganize repo for docker distribution
