+++ .gitignore
@@ -1,10 +1,6 @@
-/config.yml
-/dist/config.yml
-/venv/
-/.gitignore
-/.idea
+/config/config.yml
+/dist/plex_auto_collections
 /build
-/plex_auto_collections.spec
+/app/plex_auto_collections.spec
 /__pycache__/
-/config-*
 /.vscode
